# Simplify `cholesky` for `ScalMat` and `PDiagMat`