--- a/src/pdiagmat.jl
+++ b/src/pdiagmat.jl
@@ -25,7 +25,7 @@ Base.convert(::Type{AbstractPDMat{T}}, a::PDiagMat) where {T<:Real} = convert(PD
 Base.size(a::PDiagMat) = (a.dim, a.dim)
 Base.Matrix(a::PDiagMat) = Matrix(Diagonal(a.diag))
 LinearAlgebra.diag(a::PDiagMat) = copy(a.diag)
-LinearAlgebra.cholesky(a::PDiagMat) = cholesky(Diagonal(a.diag))
+LinearAlgebra.cholesky(a::PDiagMat) = Cholesky(Diagonal(map(sqrt, a.diag)), 'U', 0)
 
 ### Inheriting from AbstractMatrix