--- a/test/testutils.jl
+++ b/test/testutils.jl
@@ -120,12 +120,17 @@ function pdtest_diag(C, Cmat::Matrix, cmat_eq::Bool, verbose::Int)
     end
 end
 
-function pdtest_cholesky(C::Union{PDMat, PDiagMat}, Cmat::Matrix, cmat_eq::Bool, verbose::Int)
+function pdtest_cholesky(C::Union{PDMat, PDiagMat, ScalMat}, Cmat::Matrix, cmat_eq::Bool, verbose::Int)
     _pdt(verbose, "cholesky")
     if cmat_eq
         @test cholesky(C).U == cholesky(Cmat).U
     else
         @test cholesky(C).U ≈ cholesky(Cmat).U
+    end
+    # regression test: https://github.com/JuliaStats/PDMats.jl/pull/182
+    if C isa Union{PDiagMat, ScalMat}
+        size_of_sqrt_diag = C.dim * sizeof(float(eltype(C)))
+        @test (@allocated cholesky(C)) <= 1.05 * size_of_sqrt_diag # allow some overhead
     end
 end