@chriselrod 2023-09-26
Can you use `FillArrays.jl` to avoid the allocations?