@devmotion 2023-10-03
Based on the timings in https://github.com/JuliaStats/PDMats.jl/pull/182/#issuecomment-1736061820, I think we shouldn't switch to FillArrays in this PR. The PR is already an improvement over the status quo, and we could track the remaining possible optimizations and discussion about FillArrays in a separate issue. What do you think @chriselrod?